fixed some xpressions and the parsers is almost finished

+++ project/src/zinara/ast/expression/FloatExp.java
@@ -2,5 +2,5 @@ package zinara.ast.expression;
 
 public class FloatExp extends Expression {
    public float value;
-   public FloatExp ( int n ) { value=n; }
+   public FloatExp ( float n ) { value=n; }
 }
